requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.1.5
django-active-link==0.1.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.16.0
sqlparse==0.4.3
